+++ minimal_uhdless_demo/send_main.cpp
@@ -28,7 +28,8 @@ int main() {
     // TODO: set gain and frequency
 
     std::vector<size_t> channels = {0};
-    int64_t max_samples_per_packet = /*8912 is the maximum bytes samples can take up. sc16 are 4 bytes each*/ 8912 / 4;
+    // TODO: adjust limit for Crimson
+    size_t max_samples_per_packet = /*8912 is the maximum bytes samples can take up. sc16 are 4 bytes each*/ 8912 / 4;
     int64_t buffer_size = (int64_t) management_iface.get_double("system/get_max_buffer_level");
     std::vector<std::string> dst_ips = { "10.10.10.2" };
     std::vector<int> dst_ports = { 42836 };
@@ -37,7 +38,7 @@ int main() {
     // Packets must be a multiple of a certain length on Cyan depending on the FPGA version
     // You can get it dynamically on Cyan
     // Check CRIMSON_TNG_PACKET_NSAMP_MULTIPLE in https://github.com/pervices/uhd if porting this to Crimson
-    int packet_size_multiple = management_iface.get_int("system/nsamps_multiple_tx");
+    size_t packet_size_multiple = static_cast<size_t>(management_iface.get_int("system/nsamps_multiple_tx"));
 
     // Calculated the time of the device a packet sent now would arrive at
     std::shared_ptr<clock_sync> clock_sync_manager =  clock_sync::make("10.10.10.2", 42809, tick_rate);
